Massive Refactoring
Library now has methods for retrieving specific Song, Album, and Artist 
objects by their defining values. All songs will now be part of an 
album, and all albums will now be part of an artist, and artists can no 
longer hold songs directly. This is made possible by one blank-named 
artist, and one blank-named album per artist, including the blank-named 
one. The getAlbum command also deals with adding albums to artists, so a 
song only needs to ensure it is given to the proper album to be in the 
right place in the chain